Final project submission
- Created app to have search field ALWAYS visible (unless the URL page
cannot be found).  There was a little confusion as the project
instructions state one thing, while the photo layout shows another.
The project instructions state, “Add search field feature. The field
should be visible only in the 'Search' route.”  I ignored this and
chose to design the app to look like the photo layout.